Edited README
These are reflective of changes I'm doing across multiformats; some of them are stylistic, and some of them reflect the need for a few other things - a Code of Conduct, for instance. Install and Usage need more work; I'll open an issue for those. As well, it would be great to have code coverage and ci. But that is outside of the scope of this.

For now:

- Fix https issues for project and freenode badges.
- Added standard-readme badge.
- Changed `Maintainer` to `Maintainers` to match standard-readme.
- Added Install and Usage sections with TODOs.
- Fixed spelling error for contributions.
- Edited license text to match standard-readme.
- Added Contributing document for organization
- Added note about standard readme.
- Added link to IPFS Code of Conduct.